lib: mpv_player: Implement playlist monitoring
Since we need to provide proper error messages when MPV throws errors
on loading files, we need to be in sync with MPV events, and monitoring
from the UI code can be convoluted. This commit moves the monitoring to
the MPV-related code, where proper event logging and error handling can
be done.