ui: mpv_player: Do not retry connection to socket on closing instances
Previously, CloseInstances would continually loop trying to connect to a
socket, to send a quit signal to the mpv server. This is redundant, since
it can keep retrying until the max retry limit has been reached, especially
in the case of a socket that is inactive/has no listeners. So we check for
open connections only once, close any active mpv servers, and proceed.