lib: video: Use FormatStreams to check itags
Previously, for videos a video url was formed based on the assumption
that certain itags existed. For example, for the 720p stream video,
a videoUrl would be returned directly with the corresponding itag,
without checking whether it exists, or it really is the correct itag.
FormatStreams contains the correct data that we can use to check and
return a video url, so use it instead.