ui: player: Ensure player state and history is saved only on user inititated exit
At certain times, the player state and history is saved erroneously when
the following scenarios occurs in sequence:

- A fresh instance (instanceA) is launched normally, tracks are added to the
  playlist queue and player settings (mute, shuffle, loop) are changed.
  This instance is suspended to shell.

- An alternate instance (instanceB) is launched with the "--close-instances" option,
  different tracks are added to the playlist queue, and player setting are changed.
  This instance is quit normally by the user.

- The suspended instance "instanceA" is brought to the foreground. Since "instanceB"
  has already quit, "instanceA" will automatically shutdown.

When this scenario occurs, the state and history data of "instanceA" is saved,
despite the fact that "instanceB" underwent a user-initiated exit, and therefore
"instanceB" should have its player state and history data saved.

Detect when an instance undergoes a user inititated exit, and only then save the player data.