ui: playlist: Optimize playlist updation
With this commit, the playlist is only updated when changes are
detected in the playlist data (MPV's playlist event) and when
the playlist popup is visible. This saves a considerable amount of
CPU cycles, especially since we don't have to use a timer to sync
with MPV's playlist, and we currently don't have to compare slices
to detect changes before a redraw.

Note that CPU usage spikes can happen when the playlist popup is open
and multiple playlist-related events are happening, such as
when a track is changed, added or removed. For large playlists,
this will still take somewhat more CPU cycles, due to redraws
happening per event. This will be addressed in another follow-up commit.